requirements.txt
numpy
scipy
open3d>=0.17.0
trimesh
scikit-learn
imageio
opencv-python
PyYAML
tqdm
pynvx
